refactor

+++ src/main/java/io/tatum/model/response/bch/BchTxVin.java
@@ -17,7 +17,7 @@ import java.math.BigDecimal;
 public class BchTxVin {
 
     private String txid;
-    private BigDecimal vout;
+    private int vout;
     private BchTxScriptSig scriptSig;
     private String coinbase;
     private long sequence;
